agent/pyproject.toml
[project]
name = "nexus-agent"
version = "0.1.0"
description = "NEXUS - AI agent with full Linux desktop control"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "e2b-desktop>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.8.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nexus*"]
